nop bai tap 18/04

+++ src/ss11_map_and_tree/bai_tap/luyen_tap_arraylist_linkedlist/model/Product.java
@@ -3,9 +3,9 @@ package ss11_map_and_tree.bai_tap.luyen_tap_arraylist_linkedlist.model;
 public class Product {
     private int id;
     private String name;
-    private int price;
+    private double price;
 
-    public Product(int id, String name, int price) {
+    public Product(int id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -27,11 +27,11 @@ public class Product {
         this.name = name;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return this.price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
